airflow/requirements.txt
apache-airflow-providers-google==10.13.1
requests==2.31.0
pandas==1.5.3
